backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
motor==3.3.2
pydantic==2.5.0
pillow==10.1.0
numpy<2.0.0
opencv-python-headless==4.8.1.78
pyzbar==0.1.9
fastapi-cors==0.0.6